RC_backend_api/requirements/requirements-base.txt
-i https://pypi.org/simple
asgiref==3.9.1; python_version >= '3.9'
dj-database-url==3.0.1
django==5.2.6; python_version >= '3.10'
djangorestframework==3.16.1; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'
packaging==25.0; python_version >= '3.8'
pillow==11.3.0; python_version >= '3.9'
psycopg[binary]==3.2.10; python_version >= '3.8'
psycopg-binary==3.2.10; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.15.0; python_version >= '3.9'
django-filter==25.2;python_version >='3.9'